+++ src/test/java/utils/StartMining.java
@@ -10,7 +10,7 @@ public class StartMining {
   }
   
   public static void startMining() {
-    int numThreads = 6;
+    long numThreads = 6;
     //TestUtils.getWalletRpc().startMining(numThreads, false, true);
     TestUtils.getDaemonRpc().startMining("56SWsnhejUTbgNs2EgyXdfNXUawymMMuAC9voZZSQrHzJHNxGsAvMnoUja7JcKVtPwNc1oKAkoAt1cv6EmtKRQ22U37B7cT", numThreads, false, false);  // random subaddress
   }
